+++ nppm/lib/config.py
@@ -18,122 +18,141 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-import appdirs
+from six.moves import configparser
 import errno
+import getpass
+import io
 import os
+import re
 import sys
-
-import { get_python_install_type } from './env'
 
 NPPM_CONFIG = '~/.nppmrc'
 
 
+class NoSuchSection(KeyError):
+  pass
+
+
 class Config(object):
   """
-  Reader/writer for the NPPM configuration file. The file format must be in
-  plain `key=value` format, one per line. Lines starting with a hash (`#`)
-  will be ignored.
-
-  # Parameters
-  filename (str): The config file to parse, and optionally save to. Pass
-      #NotImplemented to prevent the #Config object from actually loading any
-      configuration file. Defaults to the value of the `NPPM_CONFIG` environment
-      variable or `~/.nppmrc`.
-  defaults (dict): A dictionary of default values.
+  Wraps a #configparser.SafeConfigParser object and normalizes registry
+  section names as well as creating a default `__global__` section at the
+  start.
   """
 
-  def __init__(self, filename=NotImplemented, defaults=None):
-    if filename is NotImplemented:
-      filename = os.getenv('NPPM_CONFIG', NPPM_CONFIG)
-    if filename:
-      filename = os.path.normpath(os.path.expanduser(filename))
-    if defaults is None:
-      defaults = {}
+  NoSuchSection = NoSuchSection
 
+  def __init__(self, filename, defaults=None):
     self.filename = filename
-    self.defaults = defaults
-    self._values = {}
-    self._values_ordered = []
+    self._parser = configparser.SafeConfigParser()
+    self.defaults = defaults or {}
+    self.read(filename, doraise=False)
 
-    if self.filename is not None:
-      self.load()
-
-  def __repr__(self):
-    return '<Config {!r}>'.format(self.filename)
-
-  def load(self, filename=None, ignore_not_exists=True):
-    """
-    Loads the configuration from the specified *filename*. Defaults to the
-    *filename* specified when the #Config object was created.
-    """
-
-    if filename is None:
-      if self.filename is None:
-        raise ValueError('no filename specified')
-
-    try:
-      with open(self.filename, 'r') as fp:
-        for line in fp:
-          line = line.rstrip('\n')
-          if line.startswith('#') or not line:
-            self._values_ordered.append([None, line])
-          else:
-            key, value = line.rstrip('\n').partition('=')[::2]
-            key = key.lower().strip()
-            value = (value or '').strip()
-            self._values[key] = value
-            self._values_ordered.append([key, value])
-    except (OSError, IOError) as exc:
-      if exc.errno != errno.ENOENT or not ignore_not_exists:
-        raise
-
-  def save(self, filename=None, fileobj=None):
-    """
-    Saves the configuration to the specified *filename*. Defaults to the
-    *filename* specified when the #Config object was created.
-    """
-
-    if filename is None and fileobj is None:
-      if self.filename is None:
-        raise ValueError('no filename specified')
-      filename = self.filename
-
-    if not self._values_ordered:
+  def read(self, filename, doraise=True):
+    if not doraise and not os.path.isfile(filename):
       return
+    with open(filename) as fp:
+      return self.readfp(fp)
 
-    # Make sure the directory we write to exists.
-    dirname = os.path.dirname(self.filename)
-    if not os.path.isdir(dirname):
-      os.makedirs(dirname)
+  def readfp(self, fp):
+    fp = io.StringIO('[__global__]\n' + fp.read())
+    self._parser.readfp(fp)
 
-    try:
-      if fileobj is None:
-        fileobj = open(filename, 'w')
-
-      for key, value in self._values_ordered:
-        if key is None:
-          fileobj.write(value)
-        else:
-          fileobj.write('{}={}'.format(key, value))
-        fileobj.write('\n')
-    finally:
-      if filename is not None:
-        fileobj.close()
+  def save(self):
+    buf = io.StringIO()
+    self._parser.write(buf)
+    buf.seek(0)
+    assert buf.readline() == '[__global__]\n'
+    with open(self.filename, 'w') as dst:
+      dst.write(buf.read())
 
   def __getitem__(self, key):
-    try:
-      return self._values[key]
-    except KeyError:
-      return self.defaults[key]
+    section, sep, valuename = key.partition('.')
+    if not valuename:
+      section, valuename = '', section
+    if not section and sep:
+      raise KeyError(key)
+    if not section:
+      section = '__global__'
+    if not self._parser.has_section(section):
+      if section not in self.defaults:
+        raise KeyError(key)
+      return self.defaults[section][valuename]
+    if not self._parser.has_option(section, valuename):
+      if section in self.defaults:
+        return self.defaults[section][valuename]
+      raise KeyError(key)
+    return self._parser.get(section, valuename)
 
   def __setitem__(self, key, value):
-    self._values[key] = str(value)
-    for item in self._values_ordered:
-      if item[0] == key:
-        item[1] = value
-        break
-    else:
-      self._values_ordered.append([key, value])
+    section, sep, valuename = key.partition('.')
+    if not section and sep:
+      raise KeyError(key)
+    if not section:
+      section = '__global__'
+    if not self._parser.has_section(section):
+      self._parser.add_section(section)
+    self._parser.set(section, valuename, value)
+
+  def get(self, key, default=None):
+    try:
+      return self[key]
+    except KeyError:
+      return default
+
+  def rename_section(self, section, new_name):
+    items = self._parser.items(section)
+    self._parser.remove_section(section)
+    self._parser.add_section(new_name)
+    for key, value in items:
+      self._parser.set(new_name, key, value)
+
+  def section(self, name):
+    if not self.has_section(name):
+      raise NoSuchSection(name)
+    return SectionView(self, name)
+
+  def has_section(self, name):
+    if self._parser.has_section(name):
+      return True
+    return name in self.defaults
+
+  def registry(self, name):
+    reg = 'registry:' + name
+    if not self.has_section(reg):
+      raise NoSuchSection(reg)
+    return SectionView(self, name, reg)
+
+  def registries(self):
+    result = []
+    default_found = False
+    for section in self._parser.sections():
+      if section.startswith('registry:'):
+        name = section[9:]
+        view = self.registry(name)
+        if name == 'default':
+          default_found = True
+        result.append(view)
+    if not default_found:
+      result.insert(0, self.registry('default'))
+    return result
+
+
+class SectionView(object):
+
+  def __init__(self, config, name, section_prefix=None):
+    self.config = config
+    self.name = name
+    self.section_prefix = section_prefix or name
+
+  def __repr__(self):
+    return '<SectionView {!r}>'.format(self.section_prefix)
+
+  def __getitem__(self, key):
+    return self.config[self.section_prefix + '.' + key]
+
+  def __setitem__(self, key, value):
+    self.config[self.section_prefix + '.' + key] = value
 
   def get(self, key, default=None):
     try:
@@ -142,10 +161,17 @@ class Config(object):
       return default
 
 
-config = Config()
+filename = os.getenv('NPPM_CONFIG', NPPM_CONFIG)
+filename = os.path.normpath(os.path.expanduser(filename))
 
-# The URL of the NPPM registry from which packages should be downloaded
-# from and uploaded to.
-config.defaults['registry'] = 'https://registry.nodepy.org'
+config = Config(filename, {
+  '__global__': {
+    'author': getpass.getuser(),
+    'license': 'MIT'
+  },
+  'registry:default': {
+    'url': 'https://registry.nodepy.org'
+  }
+})
 
 exports = config